requirements.txt
Flask==1.1.2
Requests
google-cloud-storage
